Improved code

+++ src/assetHandling/Font.cc
@@ -34,7 +34,7 @@ Font FontLoadFromFile(SDL_Renderer *renderer, const char *path) {
 }
 
 void RenderChar(SDL_Renderer *renderer, const Font *font, char c,
-                Vector2<float> pos, float scale) {
+                Vector2<int> pos, float scale) {
     const SDL_Rect dst = {
         .x = (int)floorf(pos.x),
         .y = (int)floorf(pos.y),
@@ -50,11 +50,11 @@ void RenderChar(SDL_Renderer *renderer, const Font *font, char c,
 }
 
 void RenderTextSized(SDL_Renderer *renderer, Font *font, const char *text,
-                     size_t text_size, Vector2<float> pos, SDL_Color color,
+                     size_t text_size, Vector2<int> pos, SDL_Color color,
                      float scale) {
     SetTextureColor(font->spritesheet, color);
 
-    Vector2<float> pen = pos;
+    Vector2<int> pen = pos;
     for (size_t i = 0; i < text_size; ++i) {
         RenderChar(renderer, font, text[i], pen, scale);
         pen.x += FONT_CHAR_WIDTH * scale;
@@ -62,6 +62,6 @@ void RenderTextSized(SDL_Renderer *renderer, Font *font, const char *text,
 }
 
 void RenderText(SDL_Renderer *renderer, Font *font, const char *text,
-                Vector2<float> pos, SDL_Color color, float scale) {
+                Vector2<int> pos, SDL_Color color, float scale) {
     RenderTextSized(renderer, font, text, strlen(text), pos, color, scale);
 }
